+++ model_utils/tests/fields.py
@@ -1,30 +1,34 @@
-import json
-
 from django.db import models
-from django.core.serializers.json import DjangoJSONEncoder
+
+try:
+    unicode()
+    str_class = basestring
+except NameError:
+    str_class = str
+
+def with_metaclass(meta, base=object):
+    return meta("NewBase", (base,), {})
 
 
-class SimpleJSONField(models.TextField):
-
-    __metaclass__ = models.SubfieldBase
+class MutableField(with_metaclass(models.SubfieldBase, models.TextField)):
 
     def to_python(self, value):
-        if value == "":
+        if value == '':
             return None
 
         try:
-            if isinstance(value, basestring):
-                return json.loads(value)
+            if isinstance(value, str_class):
+                return [int(i) for i in value.split(',')]
         except ValueError:
             pass
 
         return value
 
     def get_db_prep_save(self, value, connection):
-        if value == "":
-            return None
+        if value is None:
+            return ''
 
-        if isinstance(value, dict):
-            value = json.dumps(value, cls=DjangoJSONEncoder)
+        if isinstance(value, list):
+            value = ','.join((str(i) for i in value))
 
-        return super(SimpleJSONField, self).get_db_prep_save(value, connection)
+        return super(MutableField, self).get_db_prep_save(value, connection)