+++ model_utils/tests/models.py
@@ -5,7 +5,7 @@ from model_utils.models import TimeStampedModel, StatusModel, TimeFramedModel
 from model_utils.tracker import FieldTracker, ModelTracker
 from model_utils.managers import QueryManager, InheritanceManager, PassThroughManager
 from model_utils.fields import SplitField, MonitorField, StatusField
-from model_utils.tests.fields import SimpleJSONField
+from model_utils.tests.fields import MutableField
 from model_utils import Choices
 
 
@@ -235,6 +235,7 @@ class Spot(models.Model):
 class Tracked(models.Model):
     name = models.CharField(max_length=20)
     number = models.IntegerField()
+    mutable = MutableField()
 
     tracker = FieldTracker()
 
@@ -272,18 +273,10 @@ class TrackedMultiple(models.Model):
     number_tracker = FieldTracker(fields=['number'])
 
 
-class TrackedWithJsonField(models.Model):
-    name = models.CharField(max_length=20)
-    number = models.IntegerField()
-
-    props = SimpleJSONField()
-
-    tracker = FieldTracker()
-
-
 class ModelTracked(models.Model):
     name = models.CharField(max_length=20)
     number = models.IntegerField()
+    mutable = MutableField()
 
     tracker = ModelTracker()